+++ src/game/tests.rs
@@ -749,6 +749,7 @@ mod vintage_cube_adventure;
 mod vintage_cube_arrival;
 mod vintage_cube_artifacts;
 mod vintage_cube_creatures;
+mod vintage_cube_eldrazi;
 mod vintage_cube_graveyard;
 mod vintage_cube_hand_attack;
 mod vintage_cube_lands;